refactor(LIVE-34469): tidy up hedera integ tests
docs(LIVE-34469): spec for mocking hedera bridge in integration tests

refactor

refactor
