+++ apps/ledger-live-mobile/package.json
@@ -292,7 +292,6 @@
     "@ledgerhq/live-e2e-shared": "workspace:^",
     "@ledgerhq/speculos-transport": "workspace:^",
     "@ledgerhq/coin-cardano": "workspace:^",
-    "@ledgerhq/coin-hedera": "workspace:^",
     "@ledgerhq/test-utils": "workspace:^",
     "@react-native-community/cli": "catalog:",
     "@react-native-community/cli-platform-android": "catalog:",
